Extract insertSeriesInfo
