+++ lib/index.js
@@ -1,7 +1,7 @@
 import getPosts from './getPosts.js';
 import groupBy from './groupBy.js';
-import { ElementInsertionPosition, insertNodeRelativeToElement } from './insertNodeRelativeToElement.js';
-import times from './times.js';
+import { ElementInsertionPosition } from './insertNodeRelativeToElement.js';
+import insertSeriesInfo from './insertSeriesInfo.js';
 
 export { ElementInsertionPosition } from './insertNodeRelativeToElement.js';
 
@@ -33,14 +33,7 @@ export { ElementInsertionPosition } from './insertNodeRelativeToElement.js';
  * @property {(string|string[])} seriesTagSlugs One or more tag slugs for the serie(s) to build info for. 
  * @property {boolean} hideSinglePostSeries A flag indicating whether or not to hide series with a single post.
  * @property {string} comingSoonText The text to display to indicate more posts are coming for a single post series.
- * @property {Array<ElementInsertionInfo>} insertions An array element insersion information structures.
- */
-
-/**
- * 
- * @typedef {Object} ElementInsertionInfo
- * @property {!string} selector A CSS selector string for an element to insert relative to.
- * @property {!ElementInsertionPosition} position The position to insert a node relative to an element.
+ * @property {Array<import("./insertSeriesInfo").ElementInsertionInfo>} insertions An array element insersion information structures.
  */
 
 
@@ -211,29 +204,6 @@ export function SeriesDisplay(ghostContentApi) {
         } else {
             document.addEventListener('DOMContentLoaded', insertClosure);
         }
-    }
-
-    function insertSeriesInfo(document, seriesInfoFragment, insertions) {
-        // Insert the series info fragment into the appropriate location(s) of the document.
-        let fragments = [ seriesInfoFragment ];
-        const delayedActions = [];
-
-        for (const insertion of insertions) {
-            const elements = document.querySelectorAll(insertion.selector);
-            
-            // We need to clone the fragment if we have more than one insertion and/or element matching the selector.
-            const fragmentCloneCount = elements.length - fragments.length;
-            fragments = [...fragments, ...times(fragmentCloneCount, () => seriesInfoFragment.cloneNode(true)) ];
-            
-            for (const element of elements) {
-                const fragment = fragments.pop();
-                delayedActions.push(() => insertNodeRelativeToElement(fragment, element, insertion.position));
-            }
-        }
-
-        // We delayed the actual inserts because after a DocumentFragment is added to a document,
-        // its nodes are moved. Therefore, cloning it later means we clone an empty fragment.
-        delayedActions.forEach(action => action());
     }
